refactor(chatform): move history UI from ChatForm to GenericChatForm
